Restructure info collection
