+++ lang/query/src/info/mod.rs
@@ -1,0 +1,8 @@
+//! This module provides utilities which are used by the language
+//! server for the type-on-hover and code-action features.
+
+mod data;
+mod collect;
+
+pub use data::*;
+pub use collect::*;
